refactor ui trace graph helpers
